ECER-2662: Application repository refactor changed to partial class
