+++ src/ECER.Resources.Documents/Applications/ApplicationRepository.cs
@@ -1,34 +1,21 @@
 ﻿using AutoMapper;
-using ECER.Resources.Documents.Applications.ChildrenServices;
 using ECER.Resources.Documents.PortalInvitations;
 using ECER.Utilities.DataverseSdk.Model;
 using Microsoft.Xrm.Sdk.Client;
 
 namespace ECER.Resources.Documents.Applications;
 
-internal sealed class ApplicationRepository : IApplicationRepository
+internal sealed partial class ApplicationRepository : IApplicationRepository
 {
   private readonly EcerContext context;
   private readonly IMapper mapper;
-  private readonly IApplicationChildService<ecer_ProfessionalDevelopment> professionalDevelopmentService;
-  private readonly IApplicationChildService<ecer_WorkExperienceRef> workExperienceService;
-  private readonly IApplicationChildService<ecer_CharacterReference> characterReferenceService;
-  private readonly IApplicationChildService<ecer_Transcript> transcriptService;
 
   public ApplicationRepository(
        EcerContext context,
-       IMapper mapper,
-       IApplicationChildService<ecer_ProfessionalDevelopment> professionalDevelopmentService,
-       IApplicationChildService<ecer_WorkExperienceRef> workExperienceService,
-       IApplicationChildService<ecer_CharacterReference> characterReferenceService,
-       IApplicationChildService<ecer_Transcript> transcriptService)
+       IMapper mapper)
   {
     this.context = context;
     this.mapper = mapper;
-    this.professionalDevelopmentService = professionalDevelopmentService;
-    this.workExperienceService = workExperienceService;
-    this.characterReferenceService = characterReferenceService;
-    this.transcriptService = transcriptService;
   }
 
   public async Task<IEnumerable<Application>> Query(ApplicationQuery query, CancellationToken cancellationToken)
@@ -85,10 +72,10 @@ internal sealed class ApplicationRepository : IApplicationRepository
       context.Attach(ecerApplication);
       context.UpdateObject(ecerApplication);
     }
-    await professionalDevelopmentService.Update(ecerApplication, ecerProfessionalDevelopments);
-    await workExperienceService.Update(ecerApplication, ecerWorkExperienceReferences);
-    await characterReferenceService.Update(ecerApplication, ecerCharacterReferences);
-    await transcriptService.Update(ecerApplication, ecerTranscripts);
+    await UpdateProfessionalDevelopments(ecerApplication, ecerProfessionalDevelopments);
+    await UpdateWorkExperienceReferences(ecerApplication, ecerWorkExperienceReferences);
+    await UpdateCharacterReferences(ecerApplication, ecerCharacterReferences);
+    await UpdateTranscripts(ecerApplication, ecerTranscripts);
 
     context.SaveChanges();
     return ecerApplication.ecer_ApplicationId.Value.ToString();
@@ -139,210 +126,5 @@ internal sealed class ApplicationRepository : IApplicationRepository
       InviteType.WorkExperienceReference => await OptOutWorkExperienceReference(request),
       _ => throw new NotSupportedException($"{request.GetType().Name} is not supported")
     };
-  }
-
-  #region implementationDetails
-
-  private async Task<string> SubmitCharacterReference(string characterReferenceId, CharacterReferenceSubmissionRequest request)
-  {
-    await Task.CompletedTask;
-    var characterReference = context.ecer_CharacterReferenceSet.Single(c => c.ecer_CharacterReferenceId == Guid.Parse(characterReferenceId));
-
-    mapper.Map(request, characterReference);
-    bool certificateProvinceIdIsGuid = Guid.TryParse(request.ReferenceContactInformation.CertificateProvinceId, out Guid certificateProvinceId);
-
-    if (certificateProvinceIdIsGuid)
-    {
-      var province = context.ecer_ProvinceSet.SingleOrDefault(p => p.ecer_ProvinceId == certificateProvinceId);
-      if (province != null)
-      {
-        context.AddLink(characterReference, ecer_CharacterReference.Fields.ecer_characterreference_RefCertifiedProvinceId, province);
-      }
-    }
-    characterReference.StatusCode = ecer_CharacterReference_StatusCode.Submitted;
-    context.UpdateObject(characterReference);
-    context.SaveChanges();
-    return characterReference.ecer_CharacterReferenceId.ToString()!;
-  }
-
-  private async Task<string> SubmitWorkexperienceReference(string workExperienceReferenceId, WorkExperienceReferenceSubmissionRequest request)
-  {
-    await Task.CompletedTask;
-    var workExperienceReference = context.ecer_WorkExperienceRefSet.Single(c => c.ecer_WorkExperienceRefId == Guid.Parse(workExperienceReferenceId));
-
-    mapper.Map(request, workExperienceReference);
-    bool certificateProvinceIdIsGuid = Guid.TryParse(request.ReferenceContactInformation.CertificateProvinceId, out Guid certificateProvinceId);
-    if (certificateProvinceIdIsGuid)
-    {
-      var province = context.ecer_ProvinceSet.SingleOrDefault(p => p.ecer_ProvinceId == certificateProvinceId);
-      if (province != null)
-      {
-        context.AddLink(workExperienceReference, ecer_WorkExperienceRef.Fields.ecer_workexperienceref_RefCertifiedProvinceId, province);
-      }
-    }
-    workExperienceReference.StatusCode = ecer_WorkExperienceRef_StatusCode.Submitted;
-    context.UpdateObject(workExperienceReference);
-    context.SaveChanges();
-    return workExperienceReference.ecer_WorkExperienceRefId.ToString()!;
-  }
-
-  private async Task<string> OptOutCharacterReference(OptOutReferenceRequest request)
-  {
-    await Task.CompletedTask;
-    var characterReference = context.ecer_CharacterReferenceSet.Single(c => c.ecer_CharacterReferenceId == Guid.Parse(request.PortalInvitation!.CharacterReferenceId!));
-
-    mapper.Map(request, characterReference);
-    characterReference.ecer_WillProvideReference = ecer_YesNoNull.No;
-    characterReference.StatusCode = ecer_CharacterReference_StatusCode.Rejected;
-    characterReference.StateCode = ecer_characterreference_statecode.Inactive;
-    context.UpdateObject(characterReference);
-    context.SaveChanges();
-    return characterReference.ecer_CharacterReferenceId.ToString()!;
-  }
-
-  private async Task<string> OptOutWorkExperienceReference(OptOutReferenceRequest request)
-  {
-    await Task.CompletedTask;
-    var workexperienceReference = context.ecer_WorkExperienceRefSet.Single(c => c.ecer_WorkExperienceRefId == Guid.Parse(request.PortalInvitation!.WorkexperienceReferenceId!));
-
-    mapper.Map(request, workexperienceReference);
-    workexperienceReference.ecer_WillProvideReference = ecer_YesNoNull.No;
-    workexperienceReference.StatusCode = ecer_WorkExperienceRef_StatusCode.Rejected;
-    workexperienceReference.StateCode = ecer_workexperienceref_statecode.Inactive;
-    context.UpdateObject(workexperienceReference);
-    context.SaveChanges();
-    return workexperienceReference.ecer_WorkExperienceRefId.ToString()!;
-  }
-
-  #endregion implementationDetails
-
-  public async Task<string> UpdateWorkExReferenceForSubmittedApplication(WorkExperienceReference updatedReference, string applicationId, string referenceId, string userId, CancellationToken cancellationToken)
-  {
-    await Task.CompletedTask;
-    var application = context.ecer_ApplicationSet.FirstOrDefault(
-      d => d.ecer_ApplicationId == Guid.Parse(applicationId) && d.ecer_Applicantid.Id == Guid.Parse(userId)
-      );
-    if (application == null)
-    {
-      throw new InvalidOperationException($"Application '{applicationId}' not found");
-    }
-
-    var ecerWorkExperienceReference = mapper.Map<ecer_WorkExperienceRef>(updatedReference);
-
-    var existingWorkExperiences = context.ecer_WorkExperienceRefSet.Where(t => t.ecer_Applicationid.Id == Guid.Parse(applicationId)).ToList();
-
-    bool RefIdIsGuid = Guid.TryParse(referenceId, out Guid referenceIdGuid);
-    if (RefIdIsGuid)
-    {
-      var oldReference = existingWorkExperiences.SingleOrDefault(t => t.Id == referenceIdGuid);
-      // 1. Remove existing WorkExperienceReference
-
-      if (oldReference != null)
-      {
-        context.DeleteObject(oldReference);
-      }
-      else
-      {
-        throw new InvalidOperationException($"Reference '{referenceIdGuid}' not found for application '{applicationId}'");
-      }
-    }
-
-    // 2. Add New WorkExperienceReferences
-
-    ecerWorkExperienceReference.ecer_WorkExperienceRefId = Guid.NewGuid();
-    ecerWorkExperienceReference.StatusCode = ecer_WorkExperienceRef_StatusCode.ApplicationSubmitted;
-    ecerWorkExperienceReference.ecer_IsAdditional = true;
-    context.AddObject(ecerWorkExperienceReference);
-    context.AddLink(application, ecer_Application.Fields.ecer_workexperienceref_Applicationid_ecer, ecerWorkExperienceReference);
-    context.SaveChanges();
-
-    return ecerWorkExperienceReference.ecer_WorkExperienceRefId.ToString()!;
-  }
-
-  public async Task<string> UpdateCharacterReferenceForSubmittedApplication(CharacterReference updatedReference, string applicationId, string referenceId, string userId, CancellationToken cancellationToken)
-  {
-    await Task.CompletedTask;
-    var application = context.ecer_ApplicationSet.FirstOrDefault(
-      d => d.ecer_ApplicationId == Guid.Parse(applicationId) && d.ecer_Applicantid.Id == Guid.Parse(userId)
-      );
-    if (application == null)
-    {
-      throw new InvalidOperationException($"Application '{applicationId}' not found");
-    }
-
-    var ecerCharacterReference = mapper.Map<ecer_CharacterReference>(updatedReference);
-
-    var existingCharacterReferences = context.ecer_CharacterReferenceSet.Where(t => t.ecer_Applicationid.Id == Guid.Parse(applicationId)).ToList();
-
-    bool RefIdIsGuid = Guid.TryParse(referenceId, out Guid referenceIdGuid);
-    if (RefIdIsGuid)
-    {
-      var oldReference = existingCharacterReferences.SingleOrDefault(t => t.Id == referenceIdGuid);
-
-      // 1. Remove existing Character Reference
-      if (oldReference != null)
-      {
-        context.DeleteObject(oldReference);
-      }
-      else
-      {
-        throw new InvalidOperationException($"Reference '{referenceIdGuid}' not found for application '{applicationId}'");
-      }
-    }
-
-    // 2. Add New Character Reference
-
-    ecerCharacterReference.ecer_CharacterReferenceId = Guid.NewGuid();
-    ecerCharacterReference.StatusCode = ecer_CharacterReference_StatusCode.ApplicationSubmitted;
-    ecerCharacterReference.ecer_IsAdditional = true;
-    context.AddObject(ecerCharacterReference);
-    context.AddLink(application, ecer_Application.Fields.ecer_characterreference_Applicationid, ecerCharacterReference);
-    context.SaveChanges();
-
-    return ecerCharacterReference.ecer_CharacterReferenceId.ToString()!;
-  }
-
-  public async Task<string> ResendCharacterReferenceInvite(ResendReferenceInviteRequest request, CancellationToken cancellationToken)
-  {
-    await Task.CompletedTask;
-    var characterReference = context.ecer_CharacterReferenceSet.SingleOrDefault(c => c.ecer_CharacterReferenceId == Guid.Parse(request.ReferenceId!));
-
-    if (characterReference == null)
-    {
-      throw new InvalidOperationException($"Character reference '{request.ReferenceId}' not found");
-    }
-
-    if (characterReference.StatusCode == ecer_CharacterReference_StatusCode.Rejected || characterReference.StatusCode == ecer_CharacterReference_StatusCode.Submitted)
-    {
-      throw new InvalidOperationException($"Character reference '{request.ReferenceId}' already responded");
-    }
-
-    mapper.Map(request, characterReference);
-    characterReference.ecer_InviteAgain = true;
-    context.UpdateObject(characterReference);
-    context.SaveChanges();
-
-    return characterReference.ecer_CharacterReferenceId.ToString()!;
-  }
-
-  public async Task<string> ResendWorkExperienceReferenceInvite(ResendReferenceInviteRequest request, CancellationToken cancellationToken)
-  {
-    await Task.CompletedTask;
-    var workexperienceReference = context.ecer_WorkExperienceRefSet.SingleOrDefault(c => c.ecer_WorkExperienceRefId == Guid.Parse(request.ReferenceId!));
-
-    if (workexperienceReference == null)
-    {
-      throw new InvalidOperationException($"Work experience reference '{request.ReferenceId}' not found");
-    }
-
-    if (workexperienceReference.StatusCode == ecer_WorkExperienceRef_StatusCode.Rejected || workexperienceReference.StatusCode == ecer_WorkExperienceRef_StatusCode.Submitted)
-    {
-      throw new InvalidOperationException($"Work experience reference '{request.ReferenceId}' already responded");
-    }
-
-    workexperienceReference.ecer_InviteAgain = true;
-    context.UpdateObject(workexperienceReference);
-    context.SaveChanges();
-    return workexperienceReference.ecer_WorkExperienceRefId.ToString()!;
   }
 }
