+++ src/ECER.Resources.Documents/Configurer.cs
@@ -1,10 +1,8 @@
 ﻿using ECER.Infrastructure.Common;
 using ECER.Resources.Documents.Applications;
-using ECER.Resources.Documents.Applications.ChildrenServices;
 using ECER.Resources.Documents.Certifications;
 using ECER.Resources.Documents.MetadataResources;
 using ECER.Resources.Documents.PortalInvitations;
-using ECER.Utilities.DataverseSdk.Model;
 using Microsoft.Extensions.DependencyInjection;
 using System.Diagnostics.CodeAnalysis;
 
@@ -18,9 +16,5 @@ public class Configurer : IConfigureComponents
     configurationContext.Services.AddTransient<IApplicationRepository, ApplicationRepository>();
     configurationContext.Services.AddTransient<IPortalInvitationRepository, PortalInvitationRepository>();
     configurationContext.Services.AddTransient<IMetadataResourceRepository, MetadataResourceRepository>();
-    configurationContext.Services.AddTransient<IApplicationChildService<ecer_ProfessionalDevelopment>, ProfessionalDevelopmentsService>();
-    configurationContext.Services.AddTransient<IApplicationChildService<ecer_WorkExperienceRef>, WorkExperienceReferencesService>();
-    configurationContext.Services.AddTransient<IApplicationChildService<ecer_CharacterReference>, CharacterReferencesService>();
-    configurationContext.Services.AddTransient<IApplicationChildService<ecer_Transcript>, TranscriptsService>();
   }
 }
